añado dialog de confirmación al borrar

+++ front/src/app/subcategory/components/subcategory-table/subcategory-table.component.ts
@@ -10,6 +10,7 @@ import { EditDialogComponent } from '../../../category/components/edit-dialog/ed
 import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { EditSubcategoryDialogComponent } from '../edit-subcategory-dialog/edit-subcategory-dialog.component';
+import { ConfirmDialogComponent } from '../../../admin/components/confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-subcategory-table',
@@ -23,30 +24,30 @@ export class SubcategoryTableComponent {
   id?: number
   name?: string
 
-  displayedColumns: string[] = [ 'name','star'];
-  dataSource: MatTableDataSource<Subcategory>=new MatTableDataSource<Subcategory>([]);
-  constructor(private route: ActivatedRoute, private subcategoryService: SubcategoryService, private dialog: MatDialog, private snackBar:MatSnackBar) {}
+  displayedColumns: string[] = ['name', 'star'];
+  dataSource: MatTableDataSource<Subcategory> = new MatTableDataSource<Subcategory>([]);
+  constructor(private route: ActivatedRoute, private subcategoryService: SubcategoryService, private dialog: MatDialog, private snackBar: MatSnackBar) { }
 
   @ViewChild(MatPaginator) paginator!: MatPaginator;
   @ViewChild(MatSort) sort!: MatSort;
   ngOnInit() {
 
-   console.log(this.name);
-     this.route.params
-          .pipe(
-            switchMap(params => {
-              console.log('Parámetros recibidos:', params);
-              this.id = +params['id']; // Convertir a número
-              this.name = params['name']; // Guardar nombre
-              return this.subcategoryService.getSubcategoriesFromCategory(this.id);
-            })
-          )
-          .subscribe(subcategories => {
-            console.log('Subcategorías obtenidas:', subcategories.data);
-            this.dataSource=new MatTableDataSource(subcategories.data!);
-            this.dataSource.paginator=this.paginator
-            this.dataSource.sort=this.sort
-          });
+    console.log(this.name);
+    this.route.params
+      .pipe(
+        switchMap(params => {
+          console.log('Parámetros recibidos:', params);
+          this.id = +params['id']; // Convertir a número
+          this.name = params['name']; // Guardar nombre
+          return this.subcategoryService.getSubcategoriesFromCategory(this.id);
+        })
+      )
+      .subscribe(subcategories => {
+        console.log('Subcategorías obtenidas:', subcategories.data);
+        this.dataSource = new MatTableDataSource(subcategories.data!);
+        this.dataSource.paginator = this.paginator
+        this.dataSource.sort = this.sort
+      });
   }
 
   ngAfterViewInit() {
@@ -63,38 +64,38 @@ export class SubcategoryTableComponent {
     }
   }
 
-  addSubcategory(){
-     const subcategory:SubcategoryIns={name:"", id_category:this.id!}
-        const dialog = this.dialog.open( EditDialogComponent, {
-              width: '250px',
-              enterAnimationDuration:'400ms',
-              exitAnimationDuration:'400ms',
-              data:subcategory
-            });
+  addSubcategory() {
+    const subcategory: SubcategoryIns = { name: "", id_category: this.id! }
+    const dialog = this.dialog.open(EditDialogComponent, {
+      width: '250px',
+      enterAnimationDuration: '400ms',
+      exitAnimationDuration: '400ms',
+      data: subcategory
+    });
 
-            dialog.afterClosed().subscribe(
-              (result) => {
-                if( result) {
-                    this.subcategoryService.addSubcategory(subcategory)
-                    .subscribe((resultInsert)=>{
-                      this.dataSource.data = [...this.dataSource.data, resultInsert.data];
+    dialog.afterClosed().subscribe(
+      (result) => {
+        if (result) {
+          this.subcategoryService.addSubcategory(subcategory)
+            .subscribe((resultInsert) => {
+              this.dataSource.data = [...this.dataSource.data, resultInsert.data];
 
-                      this.snackBar.open(`La Subcategoría ${subcategory.name} ha sido insertada correctamente`,'Cerrar',{
-                         duration:2000,
-                          panelClass: ['main-snackbar'],
-                          verticalPosition: 'bottom'
-                      })
-                      console.log(this.snackBar)
+              this.snackBar.open(`La Subcategoría ${subcategory.name} ha sido insertada correctamente`, 'Cerrar', {
+                duration: 2000,
+                panelClass: ['main-snackbar'],
+                verticalPosition: 'bottom'
+              })
+              console.log(this.snackBar)
 
 
 
-                    })
-                }
-              }
-            )
+            })
+        }
+      }
+    )
   }
 
-  editSubcategory(subcategory:SubcategoryIns) {
+  editSubcategory(subcategory: SubcategoryIns) {
     const dialog = this.dialog.open(EditSubcategoryDialogComponent, {
       width: '250px',
       data: { ...subcategory },
@@ -115,24 +116,35 @@ export class SubcategoryTableComponent {
             row.id === response.data.id ? response.data : row
           );
 
-        
+
           this.dataSource.data = [...this.dataSource.data];
         }
       });
     });
   }
 
-  deletSubcategory(subcategory:Subcategory){
-
+  deletSubcategory(subcategory: Subcategory) {
+    const dialog = this.dialog.open(ConfirmDialogComponent, {
+      width: '250px',
+      data: { message: `Estas seguro que quieres borrar la subcategoría ${subcategory.name}` },
+      enterAnimationDuration: '400ms',
+      exitAnimationDuration: '400ms'
+    });
+    dialog.afterClosed().subscribe((result) => {
+      if (!result) {
+        return
+      }
       this.subcategoryService.deleteSubcategory(subcategory.id!).subscribe((result) => {
 
-        this.dataSource.data=this.dataSource.data.filter((cat)=>cat.id!=subcategory.id)
-        this.snackBar.open(`La categoría ${subcategory.name} ha sido borrada correctamente`,'Cerrar',{
+        this.dataSource.data = this.dataSource.data.filter((cat) => cat.id != subcategory.id)
+        this.snackBar.open(`La categoría ${subcategory.name} ha sido borrada correctamente`, 'Cerrar', {
           duration: 3000,
           horizontalPosition: 'center',
           verticalPosition: 'bottom',
           panelClass: '.snack'
         })
-        })
-    }
+      })
+    })
+
+  }
 }
